requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Compress
Flask-Limiter
Flask-WTF
python-dotenv
Werkzeug
psycopg2-binary
gunicorn
wtforms
email_validator
Pillow
redis
Flask-Mail==0.10.0
